add- date-fns and modify the mood streak
