+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@heroicons/react": "^2.2.0",
     "@prisma/client": "^6.4.1",
+    "date-fns": "^4.1.0",
     "next": "15.2.0",
     "prisma": "^6.4.1",
     "react": "^19.0.0",
